tech-debt: extract seed-testing-events.helpers (ROK-1183)
