+++ api/scripts/seed-testing.ts
@@ -6,10 +6,10 @@ import {
   DrizzleAsyncProvider,
 } from '../src/drizzle/drizzle.module';
 import { PostgresJsDatabase } from 'drizzle-orm/postgres-js';
-import { eq } from 'drizzle-orm';
 import * as schema from '../src/drizzle/schema';
 import * as dotenv from 'dotenv';
 import { seedUsers, seedCharacters } from './seed-testing-users.helpers';
+import { seedEventSignups } from './seed-testing-events.helpers';
 import { seedThemePreferences } from './seed-testing-theme.helpers';
 import {
   seedAvailability,
@@ -40,58 +40,9 @@ async function bootstrap() {
   try {
     const createdUsers = await seedUsers(db);
     await seedCharacters(db, createdUsers);
-
-    // =====================
-    // Create Event Signups (linked to characters)
-    // =====================
-    console.log('\n📝 Creating event signups with character links...\n');
-
-    const allEvents = await db.select().from(schema.events);
-
-    // Pre-fetch all characters for our seed users
-    const allCharacters = await db.select().from(schema.characters);
-
-    for (const event of allEvents) {
-      // Sign up 3-5 random users for each event
-      const numSignups = Math.floor(Math.random() * 3) + 3;
-      const shuffledUsers = [...createdUsers].sort(() => Math.random() - 0.5);
-      const selectedUsers = shuffledUsers.slice(0, numSignups);
-
-      for (const user of selectedUsers) {
-        // Check if signup exists
-        const existingSignup = await db
-          .select()
-          .from(schema.eventSignups)
-          .where(eq(schema.eventSignups.eventId, event.id))
-          .then((rows) => rows.find((r) => r.userId === user.id));
-
-        if (!existingSignup) {
-          // Find user's character for this event's game
-          const userChar = event.gameId
-            ? allCharacters.find(
-                (c) => c.userId === user.id && c.gameId === event.gameId,
-              )
-            : undefined;
-
-          await db.insert(schema.eventSignups).values({
-            eventId: event.id,
-            userId: user.id,
-            characterId: userChar?.id ?? null,
-            confirmationStatus: userChar ? 'confirmed' : 'pending',
-          });
-          const charInfo = userChar ? ` [${userChar.name}]` : '';
-          console.log(`  ✅ ${user.username}${charInfo} → ${event.title}`);
-        }
-      }
-    }
-
-    // =====================
-    // Create Unavailability
-    // =====================
+    await seedEventSignups(db, createdUsers);
     await seedAvailability(db, createdUsers);
-
     await seedThemePreferences(db, createdUsers);
-
     await seedGameTimeSlots(db, createdUsers);
 
     console.log('\n🎉 Testing data seed complete!');
@@ -107,4 +58,4 @@ async function bootstrap() {
   process.exit(0);
 }
 
-bootstrap();
+void bootstrap();
